refactor: Fix the incorrect file name case
This file was accidentally renamed using snakecase, whereas in fact
it should have been using kebabcase to follow the convention through
the rest of the app.
