+++ app/move/fields/index.js
@@ -30,7 +30,7 @@ const people = require('./people')
 const policeNationalComputer = require('./police-national-computer')
 const policeNationalComputerUpdate = require('./police-national-computer.update')
 const prisonTransferReason = require('./prison-transfer-reason')
-const prisonTransferReasonComments = require('./prison_transfer_reason_comments')
+const prisonTransferReasonComments = require('./prison-transfer-reason-comments')
 const toLocation = require('./to-location')
 const toLocationCourtAppearance = require('./to-location-court-appearance')
 const toLocationPrisonTransfer = require('./to-location-prison-transfer')
